feature/era-glonass-integration-r2:
moved `StringyEnum` from helpers module from
__carma-era-glonass-integration__ package to __carma-utils__
