+++ carma-era-glonass-integration/src/Carma/EraGlonass/Types/EGPropulsion.hs
@@ -21,6 +21,8 @@ import           Data.Model.Types (Ident)
 import qualified Carma.Model.Engine as Engine
 import qualified Carma.Model.Engine.Persistent as EnginePersistent
 
+import           Carma.Utils.StringyEnum
+import           Carma.Utils.StringyEnum.SwaggerSchema
 import           Carma.EraGlonass.Types.Helpers
