+++ carma-era-glonass-integration/src/Carma/EraGlonass/Types/EGUpdateCallCardAcceptCode.hs
@@ -13,6 +13,8 @@ import           Data.Aeson
 import           Data.Aeson.Types (Parser, typeMismatch)
 import           Data.Swagger
 
+import           Carma.Utils.StringyEnum
+import           Carma.Utils.StringyEnum.SwaggerSchema
 import           Carma.EraGlonass.Types.Helpers
